Added schedule page
